projects/telegram_notification_service/sprint_1/tasks/task_2/solution/scheduled_checks/pyproject.toml
[tool.poetry]
name = "scheduled-checks"
version = "0.1.0"
description = ""
authors = ["Nikita Savin <klaster26@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-crontab = "^3.0.0"

[tool.poetry.scripts]
make_schedule = "scheduled_checks.scripts.scheduler:add_to_schedule"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
